examples/generative/llm/vllm/requirements.txt
vllm==0.14.0
openai==2.12.0
transformers==4.57.2
pillow>=10.0.0
